💜 Implementing the backend API to the frontend

+++ Smash_Combos.Core/Cqrs/Combos/DeleteCombo/DeleteComboRequestHandler.cs
@@ -2,9 +2,11 @@
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Smash_Combos.Core.Services;
+using Smash_Combos.Domain.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
@@ -24,16 +26,29 @@ namespace Smash_Combos.Core.Cqrs.Combos.DeleteCombo
 
         public async Task<DeleteComboResponse> Handle(DeleteComboRequest request, CancellationToken cancellationToken)
         {
-            var combo = await _dbContext.Combos.Where(combo => combo.Id == request.ComboId && combo.UserId == request.UserId).FirstOrDefaultAsync();
+            var currentUser = await _dbContext.Users.Where(user => user.Id == request.CurrentUserId).SingleOrDefaultAsync();
+
+            if (currentUser == null)
+                throw new KeyNotFoundException($"User with id {request.CurrentUserId} does not exist");
+
+            var combo = await _dbContext.Combos
+                .Include(combo => combo.User)
+                .Where(combo => combo.Id == request.ComboId).FirstOrDefaultAsync();
+
             if (combo == null)
+                throw new KeyNotFoundException($"Combo with id {request.ComboId} does not exist");
+
+            if (combo.User.Id == currentUser.Id || currentUser.UserType == UserType.Moderator || currentUser.UserType == UserType.Admin)
             {
-                return new DeleteComboResponse { Success = false };
+                _dbContext.Combos.Remove(combo);
+                await _dbContext.SaveChangesAsync(CancellationToken.None);
+
+                return new DeleteComboResponse();
             }
-
-            _dbContext.Combos.Remove(combo);
-            await _dbContext.SaveChangesAsync(CancellationToken.None);
-
-            return new DeleteComboResponse { Success = true, Combo = _mapper.Map<ComboDto>(combo) };
+            else
+            {
+                throw new SecurityException("Not authorized to delete this combo");
+            }
         }
     }
 }
